% Auto-generated: do not edit by hand
\name{toolBox2}

\alias{toolBox2}

\title{ToolBox2 component}

\description{

}

\usage{
toolBox2(id=NULL, breakpoints=NULL, component=NULL, items=NULL,
layouts=NULL, linkedId=NULL, style=NULL, title=NULL)
}

\arguments{
\item{id}{Character. }

\item{breakpoints}{Named list. }

\item{component}{}

\item{items}{Unnamed list. }

\item{layouts}{Named list. }

\item{linkedId}{Character. }

\item{style}{Named list. }

\item{title}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

